Uploaded (and finished) task 4
The objectives of this task were:

Create data tables for salvoes and salvo locations.  The key classes / tables are:
Salvo, which has an ID, a turn number, and the ID of a game player
SalvoLocation, which has a salvo ID and a grid cell label
Link the data tables so that the JSON data for salvoes and salvo locations can be accessed (/rest/gamePlayers/1/salvoes, /rest/salvoes, /rest/salvoes/1/salvoLocations, etc).
Create sample salvo and salvo location data per email.
Display where the salvoes are on game gridsfor a specific player and game.
one grid for the player with his ships marked with opponent hits
one grid for shots fired at the opponent  
the URL /web/game.html?gp=1 should display the player email and grids for game 1.