Uploaded (and finished) task 7.
Add code on the back end web server to map URLs for joining and creating games to code that updates the data tables appropriately, and returns the JSON for the game created or joined.
Add code on the front end so that the Join Game and Create Game buttons send the appropriate command to the web server.
Add code on the front end to respond to successfully joining or creating a game to switch to the grid view of that particular game.
Add code to create links to games the player has already joined.